Change repo id types

+++ withdrawal-engine-domain/src/main/java/za/co/momentum/repo/ProductTypeRepository.java
@@ -3,5 +3,5 @@ package za.co.momentum.repo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import za.co.momentum.model.ProductType;
 
-public interface ProductTypeRepository extends JpaRepository<ProductType, Long> {
+public interface ProductTypeRepository extends JpaRepository<ProductType, Integer> {
 }
